fix issues, improve test, add full implementation

+++ src/main/java/ICreditCalc.java
@@ -3,5 +3,5 @@ public interface ICreditCalc {
 
     public double totalAmountPlusPercent(Credit credit);
 
-    public double overpayment(Credit credit);
+    public int overpayment(Credit credit);
 }
